+++ API.php
@@ -12,6 +12,7 @@ namespace Piwik\Plugins\WeatherReports;
 use Piwik\Archive;
 use Piwik\DataTable;
 use Piwik\Piwik;
+use Piwik\Plugins\WeatherReports\Settings\SiteUnitsStorage;
 
 /**
  * @method static \Piwik\Plugins\WeatherReports\API getInstance()
@@ -25,11 +26,12 @@ class API extends \Piwik\Plugin\API
     public function getCondition($idSite, $period, $date, $segment = false)
     {
         $dataTable = $this->getDataTable('WeatherReports_Condition', $idSite, $period, $date, $segment);
+
         $language = Conditions::getCurrentLanguage();
         $dataTable->filter(static function (DataTable $table) use ($language) {
-
             ConditionLabels::translateTable($table, $language);
         });
+
         return $dataTable;
     }
 
@@ -83,6 +85,53 @@ class API extends \Piwik\Plugin\API
         return $this->getDataTable('WeatherReports_WindDirection', $idSite, $period, $date, $segment);
     }
 
+    /**
+     * Units in which the weather values of a site are stored and displayed.
+     *
+     * @return array<string, string> quantity => unit code
+     */
+    public function getSiteUnits($idSite)
+    {
+        Piwik::checkUserHasAdminAccess($idSite);
+
+        return Units::getUnitCodes(SiteUnitsStorage::read((int) $idSite));
+    }
+
+    /**
+     * Sets the units of a site. Values tracked with a unit system marker are converted to these units, data
+     * tracked before the change is not converted.
+     *
+     * @param int $idSite
+     * @param string $temperature c or f
+     * @param string $precipitation mm or in
+     * @param string $pressure mb or in
+     * @param string $visibility km or miles
+     * @param string $windSpeed kph or mph
+     * @return bool
+     */
+    public function setSiteUnits($idSite, $temperature, $precipitation, $pressure, $visibility, $windSpeed)
+    {
+        Piwik::checkUserHasAdminAccess($idSite);
+
+        $units = [
+            Units::TEMPERATURE => (string) $temperature,
+            Units::PRECIPITATION => (string) $precipitation,
+            Units::PRESSURE => (string) $pressure,
+            Units::VISIBILITY => (string) $visibility,
+            Units::WIND => (string) $windSpeed,
+        ];
+
+        foreach ($units as $quantity => $unit) {
+            if (!Units::isKnownUnit($quantity, $unit)) {
+                throw new \Exception(Piwik::translate('WeatherReports_InvalidUnit', [$unit, $quantity]));
+            }
+        }
+
+        SiteUnitsStorage::save((int) $idSite, $units);
+
+        return true;
+    }
+
     protected function getDataTable($name, $idSite, $period, $date, $segment)
     {
         Piwik::checkUserHasViewAccess($idSite);
@@ -101,7 +150,6 @@ class API extends \Piwik\Plugin\API
     {
         $table = $this->getDataTable($name, $idSite, $period, $date, $segment);
         $table->filter('GroupBy', ['label', static function ($label) {
-
             return is_numeric($label) ? (string) (int) round((float) $label) : $label;
         }]);
         return $table;